pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "memory-palace"
version = "1.0.0"
description = "Agent Memory System — 记忆宫殿 + 搬运小人"
requires-python = ">=3.11"

dependencies = [
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "pyyaml>=6.0",
    "structlog>=24.1",
    "typer>=0.9",
    "rich>=13.0",
    "httpx>=0.27",
    "chromadb>=1.0",
    "fastmcp>=2.0,<3.0",
]

[project.scripts]
palace = "memory_palace.integration.cli:app"
palace-mcp = "memory_palace.integration.mcp_server:main"

[project.optional-dependencies]
graph = [
    "kuzu>=0.4",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.4",
    "kuzu>=0.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "minimax: MiniMax real LLM tests (require MINIMAX_API_KEY)",
    "real_llm: tests that call real LLM APIs (not mocked)",
]

[tool.ruff]
target-version = "py311"
line-length = 99

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP"]
